Move Diagrams component to it's own module
