+++ pyproject.toml
@@ -84,7 +84,7 @@ gaphorconvert = 'gaphor.plugins.diagramexport.gaphorconvert:main'
 "sanitizer" = "gaphor.UML.sanitizerservice:SanitizerService"
 "namespace" = "gaphor.ui.namespace:Namespace"
 "toolbox" = "gaphor.ui.toolbox:Toolbox"
-"diagrams" = "gaphor.ui.mainwindow:Diagrams"
+"diagrams" = "gaphor.ui.diagrams:Diagrams"
 "consolewindow" = "gaphor.plugins.console:ConsoleWindow"
 "elementeditor" = "gaphor.ui.elementeditor:ElementEditor"
 "diagram_export" = "gaphor.plugins.diagramexport:DiagramExport"
